UserController verdeeld
Added Controllers
+ Validation
+ SignIn
+ Settings
+ ResetPassword
+ Registration

Put old Controllers in a different folder that is still being used by functions that have not been migraded

Validation Controller
+ CheckLogin
+ CheckEmail
+ EmailIsUnique
+ ScaleImage

Settings Controller
+ UpdatePassword
+ SetLoginEmail
+ GetPreferenceFromUser
+ RegisterInterestsInDatabase
+ DeleteAccount (not working)
+ Logout
+ SetPreference
+ GetPickerData

Other NEW Controllers have not been touched
