--- Controller/Authentication.cs
@@ -1,10 +1,0 @@
-using DataModel;
-
-namespace Controller;
-
-
-public static class Authentication {
-
-    public static User CurrentUser { get; set; }
-
-}
