+++ Controller/OldControllers/Authentication.cs
@@ -1,0 +1,10 @@
+using DataModel;
+
+namespace Controller;
+
+
+public static class Authentication {
+
+    public static User CurrentUser { get; set; }
+
+}
