+++ Controller/RegistrationController.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Winder.Repositories.Interfaces;
+
+namespace Controller
+{
+    public class RegistrationController
+    {
+        private readonly IUserRepository _userRepository;
+
+        public RegistrationController(IUserRepository userRepository)
+        {
+            _userRepository = userRepository;
+        }
+
+
+
+    }
+}
